pyproject.toml
[project]
name = "clawlite"
version = "0.1.0"
description = "Terminal LLM Agent for Small Local Models"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
clawlite = "clawlite.__main__:main"
